# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AAST Trainery API's",
        "description": "Trainery API documentation",
        "contact": {
            "email": "admin@aast-trainery.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Trainery Demo Server"
        }
    ],
    "paths": {
...
        "/A/student/profile/education": {
            "get": {
                "tags": [
                    "A-Student Education Info"
                ],
                "summary": "Get Student education information",
                "description": "Get Student education Info",
                "operationId": "AgetStudentEducationInfo",
                "responses": {
                    "200": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessOkVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "A-Student Education Info"
                ],
                "summary": "Create Student Education Info",
                "description": "Create Student Education Info",
                "operationId": "App\\Http\\Controllers\\API\\V1\\Mobile\\StudentEducationController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentEducationInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessAcceptedVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/A/student/profile/education/{id}": {
            "get": {
                "tags": [
                    "A-Student Education Info"
                ],
                "summary": "Get Student Education Info",
                "description": "Get Student Education Info",
                "operationId": "AgetStudentEducation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Education id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessOkVirtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response403Virtual"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response404Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "A-Student Education Info"
                ],
                "summary": "Update Student Education Info",
                "description": "Update Student Education Info",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Education id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentEducationInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessAcceptedVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "A-Student Education Info"
                ],
                "summary": "Delete Student Education Info",
                "description": "delete Student Education Info",
                "operationId": "AdeleteStudentEducation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Education id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessOkVirtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response403Virtual"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response404Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StudentEducationInfoRequest": {
                "title": "Student Login Request",
                "description": "Student Login Request body data",
                "required": [
                    "school_name",
                    "city",
                    "country",
                    "from",
                    "to",
                    "cred_url"
                ],
                "properties": {
                    "school_name": {
                        "title": "School name",
                        "description": "School name",
                        "type": "string",
                        "example": "Arab academy school"
                    },
                    "city": {
                        "title": "city",
                        "description": "School city",
                        "type": "string",
                        "example": "Cairo"
                    },
                    "country": {
                        "title": "country",
                        "description": "School country",
                        "type": "string",
                        "example": "Egypt"
                    },
                    "from": {
                        "title": "school start date",
                        "description": "school start date",
                        "type": "string",
                        "example": "2001-1-1"
                    },
                    "to": {
                        "title": "school end date",
                        "description": "school end date",
                        "type": "string",
                        "example": "2017-1-1"
                    },
                    "cred_url": {
                        "title": "School credential URL",
                        "description": "School credential URL",
                        "type": "string",
                        "example": "www.google.com"
                    },
                    "cred": {
                        "title": "School Image",
                        "description": "School Image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "SuccessAcceptedVirtual": {
                "title": "success accepted 202",
                "description": "success accepted 202",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 201
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "example": "success created"
                    },
                    "response": {
                        "title": "response",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "description": "null or data sent in the request body",
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessOkVirtual": {
                "title": "success 200",
                "description": "success 200",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 200
                    },
                    "message": {
                        "title": "message",
                        "description": "Success message",
                        "type": "string",
                        "example": "success"
                    },
                    "response": {
                        "title": "response",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "description": "null or data sent in the request body",
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Response401Virtual": {
                "title": "response Error 401",
                "description": "response Error 401",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 401
                    },
                    "message": {
                        "title": "message",
                        "description": "unauthorized",
                        "type": "string",
                        "example": "unauthorized user"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Response403Virtual": {
                "title": "response Error 403",
                "description": "response Error 403",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 403
                    },
                    "message": {
                        "title": "message",
                        "description": "forbidden",
                        "type": "string",
                        "example": "forbidden"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Response404Virtual": {
                "title": "response Error 404",
                "description": "response Error 404",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "title": "message",
                        "description": "not found",
                        "type": "string",
                        "example": "Not found"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object",
                                "example": null
                            },
                            "meta": {
                                "type": "object",
                                "example": null
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object",
                        "example": null
                    }
                },
                "type": "object"
            },
            "Response422Virtual": {
                "title": "response validation Error 422",
                "description": "response validation Error 422",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 422
                    },
                    "message": {
                        "title": "message",
                        "description": "invalid data",
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field is required"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "<b>User Token:-<b><br>\n                                        <b>Header key:</b> <i>Authorization</i><br>\n                                        <b>Scheme:</b> <i>bearer</i><br>\n                                        <b>Example:</b> <i>Authorization: Bearer {JWT_token}</i>\n                                        ",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/api/1.0/auth/login",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}